+++ R/build.R
@@ -33,7 +33,7 @@
 #' }
 drake_build <- function(
   target,
-  config = drake::read_drake_config(envir = envir, jobs = jobs),
+  config = NULL,
   meta = NULL,
   character_only = FALSE,
   envir = parent.frame(),
@@ -50,10 +50,18 @@ drake_build <- function(
   if (!character_only) {
     target <- as.character(substitute(target))
   }
+  force(envir)
+  if (is.null(config)) {
+    config <- drake::read_drake_config(envir = envir, jobs = jobs)
+    config$envir <- envir
+  }
+  # Wait until drake 7.0.0 to uncomment
+  # lock_environment(config$envir) # nolint
+  # on.exit(unlock_environment(config$envir)) # nolint
   loadd(
     list = target,
     deps = TRUE,
-    envir = envir,
+    envir = config$eval,
     cache = config$cache,
     graph = config$graph,
     jobs = jobs,