+++ R/clustermq.R
@@ -93,7 +93,7 @@ cmq_deps_list <- function(target, config) {
   out <- lapply(
     X = deps,
     FUN = function(name) {
-      config$envir[[name]]
+      config$eval[[name]]
     }
   )
   names(out) <- deps
@@ -110,13 +110,15 @@ cmq_deps_list <- function(target, config) {
 #' @param deps named list of target dependencies
 #' @param config a [drake_config()] list
 cmq_build <- function(target, meta, deps, config) {
+  # Wait until drake 7.0.0 to uncomment
+  # lock_environment(config$envir) # nolint
   if (identical(config$garbage_collection, TRUE)) {
     gc()
   }
   do_prework(config = config, verbose_packages = FALSE)
   if (identical(config$caching, "master")) {
     for (dep in names(deps)) {
-      config$envir[[dep]] <- deps[[dep]]
+      config$eval[[dep]] <- deps[[dep]]
     }
   } else {
     manage_memory(targets = target, config = config, jobs = 1)