+++ R/distributed.R
@@ -20,12 +20,6 @@ prepare_distributed <- function(config) {
   if (!file.exists(config$cache_path)) {
     dir.create(config$cache_path)
   }
-  writeLines(
-    text = as.character(config$fetch_cache),
-    con = file.path(config$cache_path, fetch_cache_file)
-  )
-  # Always save globalenv() because config$envir could inherit from it
-  # and so drake might look for stuff there.
   save(
     list = setdiff(ls(globalenv(), all.names = TRUE), config$plan$target),
     envir = globalenv(),
@@ -57,20 +51,10 @@ build_distributed <- function(target, cache_path, check = TRUE) {
 }
 
 recover_drake_config <- function(cache_path) {
-  fetch_cache <- tryCatch({
-      out <- readLines(con = file.path(cache_path, fetch_cache_file))
-      paste0(out, collapse = "\n")
-    },
-    error = error_null
-  )
-  cache <- this_cache(cache_path, verbose = FALSE, fetch_cache = fetch_cache)
+  cache <- this_cache(cache_path, verbose = FALSE)
   config <- read_drake_config(cache = cache)
-  # Always load globalenv() because config$envir could inherit from it
-  # and so drake might look for stuff there.
   dir <- cache_path(cache = cache)
   file <- globalenv_file(dir)
   load(file = file, envir = globalenv())
   config
 }
-
-fetch_cache_file <- "fetch_cache.R"