+++ R/future.R
@@ -48,8 +48,10 @@ run_future <- function(config) {
 #' @param meta list of metadata
 #' @param config [drake_config()] list
 #' @param protect Names of targets that still need their
-#' dependencies available in `config$envir`.
+#' dependencies available in memory.
 drake_future_task <- function(target, meta, config, protect) {
+  # Wait until drake 7.0.0 to uncomment
+  # lock_environment(config$envir) # nolint
   if (identical(config$caching, "worker")) {
     manage_memory(targets = target, config = config, downstream = protect)
   }