+++ R/map_reduce.R
@@ -35,16 +35,19 @@
 #' covariates <- setdiff(colnames(mtcars), "mpg")
 #' args <- t(combn(covariates, 2))
 #' colnames(args) <- c("x1", "x2")
-#' args <- tibble::as_tibble(args)
+#' # Use tibble::as_tibble(args) for better printing # nolint
+#' # Below, stringsAsFactors = FALSE is very important! # nolint
+#' args <- as.data.frame(args, stringsAsFactors = FALSE)
 #' args$data <- "mtcars"
 #' args$data <- rlang::syms(args$data)
 #' args$id <- paste0("fit_", args$x1, "_", args$x2)
-#' args
+#' # print(args) # Requires `args` to be a tibble # nolint
 #' plan <- map_plan(args, my_model_fit)
 #' plan
 #' # Consider `trace = TRUE` to include the columns in `args`
 #' # in your plan.
-#' map_plan(args, my_model_fit, trace = TRUE)
+#' plan <- map_plan(args, my_model_fit, trace = TRUE)
+#' # print(plan) # If you have tibble installed # nolint
 #' # And of course, you can execute the plan and
 #' # inspect your results.
 #' cache <- storr::storr_environment()
@@ -57,7 +60,7 @@ map_plan <- function(
   character_only = FALSE,
   trace = FALSE
 ) {
-  args <- tibble::as_tibble(args)
+  args <- weak_as_tibble(args)
   if (!character_only) {
     fun <- as.character(substitute(fun))
     id <- as.character(substitute(id))
@@ -79,9 +82,9 @@ map_plan <- function(
       wide_deparse(out)
     }
   )))
-  out <- tibble::tibble(target = target, command = command)
+  out <- weak_tibble(target = target, command = command)
   if (trace) {
-    out <- tibble::as_tibble(cbind(out, args))
+    out <- weak_as_tibble(cbind(out, args))
   }
   sanitize_plan(out)
 }
@@ -139,7 +142,7 @@ gather_plan <- function(
   command <- paste(plan$target, "=", plan$target)
   command <- paste(command, collapse = ", ")
   command <- paste0(gather, "(", command, ")")
-  new_plan <- tibble::tibble(target = target, command = command)
+  new_plan <- weak_tibble(target = target, command = command)
   if (append) {
     bind_plans(plan, new_plan)
   } else {
@@ -221,7 +224,7 @@ gather_by <- function(
     gather = gather,
     append = FALSE
   )
-  cols <- gathered[, col_names]
+  cols <- gathered[, col_names, drop = FALSE]
   suffix <- apply(X = cols, MARGIN = 1, FUN = paste, collapse = sep)
   if (length(suffix)) {
     suffix[nzchar(suffix)] <- paste0(sep, suffix[nzchar(suffix)])
@@ -302,9 +305,10 @@ reduce_plan <- function(
       base_name = paste0(target, sep)
     )
     pairs$names[nrow(pairs)] <- target
-    out <- tibble::tibble(
+    command <- paste0(begin, pairs$odds, op, pairs$evens, end)
+    out <- weak_tibble(
       target = pairs$names,
-      command = paste0(begin, pairs$odds, op, pairs$evens, end)
+      command = command[seq_len(length(pairs$names))]
     )
   } else {
     command <- Reduce(
@@ -313,7 +317,7 @@ reduce_plan <- function(
         paste0(begin, x, op, y, end)
       }
     )
-    out <- tibble::tibble(target = target, command = command)
+    out <- weak_tibble(target = target, command = command)
   }
   if (append) {
     bind_plans(plan, out)
@@ -405,7 +409,7 @@ reduce_by <- function(
     append = FALSE,
     sep = sep
   )
-  cols <- reduced[, col_names]
+  cols <- reduced[, col_names, drop = FALSE]
   suffix <- apply(X = cols, MARGIN = 1, FUN = paste, collapse = sep)
   if (length(suffix)) {
     suffix[nzchar(suffix)] <- paste0(sep, suffix[nzchar(suffix)])