+++ R/utils.R
@@ -227,8 +227,8 @@ split_by <- function(.x, .by = character(0)) {
   if (!length(.by)) {
     return(list(.x))
   }
-  f <- lapply(.x[, .by], factor, exclude = c())
-  splits <- split(x = .x, f = f)
+  fact <- lapply(.x[, .by, drop = FALSE], factor, exclude = c())
+  splits <- split(x = .x, f = fact)
   Filter(x = splits, f = nrow)
 }
 
@@ -269,10 +269,59 @@ unwrap_function <- function(funct) {
   funct
 }
 
+lock_environment <- function(envir) {
+  lockEnvironment(envir, bindings = FALSE)
+  lapply(
+    X = ls(envir, all.names = FALSE),
+    FUN = lockBinding,
+    env = envir
+  )
+}
+
+unlock_environment <- function(envir) {
+  if (is.null(envir)) {
+    stop("use of NULL environment is defunct")
+  }
+  if (!inherits(envir, "environment")) {
+    stop("not an environment")
+  }
+  .Call("unlock_environment", envir)
+  lapply(
+    X = ls(envir, all.names = FALSE),
+    FUN = unlockBinding,
+    env = envir
+  )
+  stopifnot(!environmentIsLocked(envir))
+}
+
 which_unnamed <- function(x) {
   if (!length(names(x))) {
     rep(TRUE, length(x))
   } else {
     !nzchar(names(x))
+  }
+}
+
+# weak_tibble - use tibble() if available but fall back to 
+# data.frame() if necessary
+weak_tibble <- function(..., .force_df = FALSE) {
+  no_tibble <- !suppressWarnings(requireNamespace("tibble", quietly = TRUE))
+
+  if (.force_df || no_tibble) {
+    data.frame(..., stringsAsFactors = FALSE)
+  } else {
+    tibble::tibble(...)
+  }
+}
+
+# weak_as_tibble - use as_tibble() if available but fall back to 
+# as.data.frame() if necessary
+weak_as_tibble <- function(..., .force_df = FALSE) {
+  no_tibble <- !suppressWarnings(requireNamespace("tibble", quietly = TRUE))
+
+  if (.force_df || no_tibble) {
+    as.data.frame(..., stringsAsFactors = FALSE)
+  } else {
+    tibble::as_tibble(...)
   }
 }